Move position computation to GaiaClient
