+++ src/org/helioviewer/jhv/layers/stars/GaiaClient.java
@@ -14,9 +14,12 @@ import org.helioviewer.jhv.Log;
 import org.helioviewer.jhv.astronomy.Position;
 import org.helioviewer.jhv.astronomy.Spice;
 import org.helioviewer.jhv.astronomy.SpiceMath;
+import org.helioviewer.jhv.base.Colors;
 import org.helioviewer.jhv.gui.Message;
 import org.helioviewer.jhv.io.NetClient;
 import org.helioviewer.jhv.io.UriTemplate;
+import org.helioviewer.jhv.opengl.BufVertex;
+import org.helioviewer.jhv.opengl.GLSLShape;
 import org.helioviewer.jhv.threads.EDTCallbackExecutor;
 import org.helioviewer.jhv.time.JHVTime;
 
@@ -34,25 +37,26 @@ public class GaiaClient {
         Logger.getLogger("uk.ac.starlink").setLevel(Level.WARNING); // shut-up info logs
     }
 
-    public static void submitSearch(ReceiverStars receiver, Position viewpoint) {
+    public static void submitSearch(Receiver receiver, Position viewpoint) {
         EDTCallbackExecutor.pool.submit(new QueryTap(viewpoint), new CallbackTap(receiver, viewpoint));
     }
 
-    public interface ReceiverStars {
-        void setStars(Position viewpoint, List<Star> list);
+    public interface Receiver {
+        void setStars(Position viewpoint, BufVertex pointBuf);
     }
 
     // uk.ac.starlink.table.TimeMapper.DECIMAL_YEAR.toUnixSeconds(2015.5)
     // public static final double EPOCH_2015_5_SEC = 1435838400; // DR2
     private static final double EPOCH_2016_0_SEC = 1451606400; // DR3
-    public static final double EPOCH = EPOCH_2016_0_SEC;
+    private static final double EPOCH = EPOCH_2016_0_SEC;
     private static final int SEARCH_CONE = 4;
     private static final int SEARCH_MAG = 9;
+    private static final float SIZE_POINT = 0.04f;
 
     private record StarRequest(int ra, int dec, int cone, int mag) {
     }
 
-    public record Star(int id, double ra, double dec, double pmra, double pmdec, double mag) {
+    private record Star(int id, double ra, double dec, double pmra, double pmdec, double mag) {
     }
 
     private static StarRequest computeRequest(Position viewpoint) {
@@ -63,6 +67,95 @@ public class GaiaClient {
         return new StarRequest((int) ra, (int) dec, SEARCH_CONE + 1, SEARCH_MAG);
     }
 
+    private static void putPlanet(String planet, JHVTime time, double[] sc, double[] rsc, BufVertex pointsBuf) {
+        double[] v = Spice.getPositionRec("SUN", planet, "SOLO_IAU_SUN_2009", time);
+        v[0] -= rsc[0]; // sc->planet = sun->planet - sun->sc
+        v[1] -= rsc[1];
+        v[2] -= rsc[2];
+        v = SpiceMath.recrad(v);
+
+        double[] theta = new double[2];
+        calcProj3(0, v[1], v[2], sc[1], sc[2], theta);
+        putVertex(pointsBuf, theta[0], theta[1], sc[0], 2 * SIZE_POINT, Colors.Green);
+    }
+
+    private static BufVertex computePoints(Position viewpoint, List<Star> stars) {
+        BufVertex pointsBuf = new BufVertex(500 * GLSLShape.stride);
+
+        double[] sc = new double[]{viewpoint.distance, -viewpoint.lon, viewpoint.lat}; // lon was negated
+        double[] rsc = SpiceMath.radrec(sc[0], sc[1], sc[2]);
+
+        JHVTime time = viewpoint.time;
+        putPlanet("MERCURY", time, sc, rsc, pointsBuf);
+        putPlanet("VENUS", time, sc, rsc, pointsBuf);
+        putPlanet("MARS BARYCENTER", time, sc, rsc, pointsBuf);
+
+        double dyr = (time.milli / 1000. - GaiaClient.EPOCH) / 86400. / 365.25;
+        double mas_dyr = dyr / 1000. / 3600.;
+        double[][] mat = Spice.j2000ToSun.get(time);
+
+        double[] theta = new double[2];
+        for (GaiaClient.Star star : stars) {
+            double ra = Math.toRadians(star.ra());
+            double dec = Math.toRadians(star.dec());
+
+            // http://mingus.mmto.arizona.edu/~bjw/mmt/spectro_standards.html
+            double pmra = Math.toRadians(star.pmra() * mas_dyr);
+            double pmdec = Math.toRadians(star.pmdec() * mas_dyr);
+            ra += pmra / Math.cos(dec);
+            dec += pmdec;
+
+            double[] s = SpiceMath.radrec(1, ra, dec);
+            s = SpiceMath.mxv(mat, s); // to Carrington
+            s = SpiceMath.recrad(s);
+
+            calcProj3(0, s[1], s[2], sc[1], sc[2], theta);
+            putVertex(pointsBuf, theta[0], theta[1], sc[0], 2 * SIZE_POINT, Colors.Blue);
+        }
+        return pointsBuf;
+    }
+
+    private static void putVertex(BufVertex pointsBuf, double Tx, double Ty, double dist, float size, byte[] color) {
+        double x = dist * Math.tan(Tx);
+        double y = dist * Math.tan(Ty);
+        pointsBuf.putVertex((float) x, (float) y, 0, size, color);
+    }
+
+// --Commented out by Inspection START (23/06/2024, 23:15):
+//    // https://celestialscenes.com/alma/convert/coordconvert.pdf
+//    // small angles
+//    private static void calcProj1(double P, double alpha, double delta, double alpha0, double delta0, double[] theta) {
+//        double sP = Math.sin(P);
+//        double cP = Math.cos(P);
+//        double alphaCosDelta = (alpha - alpha0) * Math.cos(delta0);
+//        theta[0] = -alphaCosDelta * cP + (delta - delta0) * sP;
+//        theta[1] = alphaCosDelta * sP + (delta - delta0) * cP;
+//    }
+// --Commented out by Inspection STOP (23/06/2024, 23:15)
+
+    // big angles
+    private static void calcProj3(double P, double alpha, double delta, double alpha0, double delta0, double[] theta) {
+        double phi = Math.atan2(Math.sin(alpha - alpha0), Math.tan(delta) * Math.cos(delta0) - Math.sin(delta0) * Math.cos(alpha - alpha0));
+        double rho = Math.acos(Math.cos(delta) * Math.cos(delta0) * Math.cos(alpha - alpha0) + Math.sin(delta) * Math.sin(delta0));
+
+        theta[0] = Math.atan(-Math.tan(rho) * Math.sin(phi - P));
+        theta[1] = Math.asin(Math.sin(rho) * Math.cos(phi - P));
+    }
+
+// --Commented out by Inspection START (23/06/2024, 23:15):
+//    // all angles
+//    private static void calcProj5(double P, double alpha, double delta, double alpha0, double delta0, double[] theta) {
+//        double phi = Math.atan2(Math.sin(alpha - alpha0), Math.tan(delta) * Math.cos(delta0) - Math.sin(delta0) * Math.cos(alpha - alpha0));
+//
+//        double num = Math.hypot(Math.cos(delta) * Math.sin(alpha - alpha0), Math.cos(delta0) * Math.sin(delta) - Math.sin(delta0) * Math.cos(delta) * Math.cos(alpha
+//        double den = Math.sin(delta) * Math.sin(delta0) + Math.cos(delta) * Math.cos(delta0) * Math.cos(alpha - alpha0);
+//        double rho = Math.atan2(num, den);
+//
+//        theta[0] = Math.atan(-num / den * Math.sin(phi - P));
+//        theta[1] = Math.asin(Math.sin(rho) * Math.cos(phi - P));
+//    }
+// --Commented out by Inspection STOP (23/06/2024, 23:15)
+
     private static final UriTemplate queryTemplate = new UriTemplate("https://gea.esac.esa.int/tap-server/tap/sync",
             UriTemplate.vars().set("REQUEST", "doQuery").set("LANG", "ADQL").set("FORMAT", "fits"));
 
@@ -71,9 +164,9 @@ public class GaiaClient {
                 String.format("1=CONTAINS(POINT('ICRS',ra,dec), CIRCLE('ICRS',%d,%d,%d)) AND phot_g_mean_mag<%d", ra, dec, cone, mag);
     }
 
-    private record QueryTap(Position viewpoint) implements Callable<List<Star>> {
+    private record QueryTap(Position viewpoint) implements Callable<BufVertex> {
         @Override
-        public List<Star> call() throws Exception {
+        public BufVertex call() throws Exception {
             // return Collections.emptyList();
             StarRequest req = computeRequest(viewpoint);
             String adql = adqlSearch(req.ra, req.dec, req.cone, req.mag);
@@ -97,14 +190,14 @@ public class GaiaClient {
                     }
                 }
                 Log.info("Found " + stars.size() + " stars with " + adql);
-                return stars;
+                return computePoints(viewpoint, stars);
             }
         }
     }
 
-    private record CallbackTap(ReceiverStars receiver, Position viewpoint) implements FutureCallback<List<Star>> {
+    private record CallbackTap(Receiver receiver, Position viewpoint) implements FutureCallback<BufVertex> {
         @Override
-        public void onSuccess(@Nonnull List<Star> result) {
+        public void onSuccess(@Nonnull BufVertex result) {
             receiver.setStars(viewpoint, result);
         }
